Prevent cstore drones from staying connected until timeout.
The delete_record subroutine was using xact_rollback and xact_commit
both of which do not close the cstore connection created by xact_begin.
This could lead to the cstore drone being unusable until the hanging
connection times out.

This commit switches to using the plain rollback and commit methods
of CStoreEditor that also close the connection if it exists.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>